Disable ansi color when either stdout or stderr isn't a tty
One of the benefits of logging is dumping it to a text file for later
review.

This change modifies loggerv so that it only prints ansi codes
when stdout and stderr are both connect to a controlling terminal.
On the other hand, when either is being redirected, color is disabled.

Ideally color would be disabled for stdout and stderr separately, but
this change is small, to the point, and takes it 90% of the way to where
it wants to be.