Change tag portion to include level
The level is added to the "tag" portion of a log statement, where the
"tag" is the text to the left of the colon, `:`. Previously, the log
statement only included in the module location and not the level. The
level was only indicated by the color, but if color was not used, then
there was no way to know the level of the log statement. Now, the level
is included and square brackets are used to delimit the module location,
i.e. `ERROR [loggerv::clap]: This is the new format for a log statement`.