Change API to be consistent
The `disable_colors` method is changed to the `colors` method with
a boolean parameter to match the rest of the methods in the Builer
pattern for the Logger struct. A check is still conducted if the logger
is used in a terminal and if it is _not_ the output is _not_ colorized
regardless of the parameter value.