Add unit tests
The `init_with_verbosity` unit test had to be removed so the `init`
method for the Builder API could be added. The `init_with_verbosity`
test felt a little out of place anyways because it was a public function
and the assertions did not check any of the private, internal values.
The remaining new tests are all units with a single assertion per test
except for the `color_works` tests as this was easier and quicker to
implement as one test with multiple assertions than multiple tests with
essentially the same assertion.