Refactor struct fields for consistency
The `module_path` and `line_numbers` fields for the `Logger` struct are
renamed to `include_module_path` and `include_line_numbers` to match the
`include_level` field, which are all booleans that indicate whether to
"include" a component in the tag portion of a log statement. The
constants defining the default values are also renamed to follow the
field names. However, the builder method names for each field do _not_
contain the "include" portion.