Add examples
The `clap` example is renamed to `quick` because multiple examples use
the clap crate and it is a "quick" demonstration of the defaults. The
clap crate uses "quick" as a name for examples as well.

The old `config` example is changed to the `run-time-config` example and
it demonstrations run-time configuration of output. The new
`compile-time-config` example is used to show how to configure the
output without using verbosity and command line arguments. The final
example, `cfg-confg`, shows how to configure the logger based on the
build profile, i.e. conditional compilation and the cfg macro.