Fix formatting of module examples
The examples in the module documentation did not match the formatting or
output in the examples for the crate. This is not a big deal, but the
comments in the last example of the module documentation were not
properly word-wrapped in the online docs. So, I took this opportunity to
remove the misaligned comments and clean-up the module examples to be
a little more consistent with the crate examples. I do not think this
requires a new release, but the docs could be updated.