Change `offset` method to `base_level`
It is a little easier to understand and example a base level than an
offset. The offset is still used internally, but now a level is used to
determine the level with a zero verbosity. This fits better with the
`max_level` method and possibly a `min_level` method in the future.