Add `output` method
This required changing the internal representation of the levels and
associated colors within the logger. The numerous `*_color` methods are
also removed. These always felt awkward. They have been replaced with
a single `color` method that uses the log level to set the color. This
same pattern is then used with the `output` method.

The `output` method allows a developer to change the standard stream,
stdout or stderr, for each level. The default is ERROR and WARN are
printed on `stderr` and everything else is printed on `stdout`. This was
hard-coded. Now, the developer can choose between the two streams on
a per-level basis. Examples and tests are also added.